pipeline/requirements.txt
dvc
openai
google-genai
python-dotenv
PyPDF2
requests
